Updated sophomore_functions.py
I realized that once you decide to remove any websites that do not contain the "usf" keyword in the link, there is no need to define a list of irrelevant keywords such as facebook, twitter, youtube, etc., since it is already removed in the first step. I think this new updated function will reduce the redundancy of the code overall. If you have any recommendations to improve or optimize the code, feel free to let me know!